Move OPA tests into OPA extension
This change moves the OPA int-tests into the OPA main module, because the reason for the separate project, the Quarkus build, no longer exists.
