+++ extensions/auth/opa/src/bearerTokenIntTest/java/org/apache/polaris/extension/auth/opa/test/OpaPolicyCatalogHandlerIT.java
