+++ extensions/auth/opa/src/main/java/org/apache/polaris/extension/auth/opa/model/OpaRequest.java
