+++ gradle/projects.main.properties
@@ -47,8 +47,7 @@ polaris-misc-types=tools/misc-types
 polaris-extensions-federation-hadoop=extensions/federation/hadoop
 polaris-extensions-federation-hive=extensions/federation/hive
 polaris-extensions-federation-bigquery=extensions/federation/bigquery
-polaris-extensions-auth-opa=extensions/auth/opa/impl
-polaris-extensions-auth-opa-tests=extensions/auth/opa/tests
+polaris-extensions-auth-opa=extensions/auth/opa
 polaris-extensions-auth-ranger=extensions/auth/ranger/impl
 
 polaris-config-docs-annotations=tools/config-docs/annotations
